+++ src/components/sections/projects/ProjectsSection/ProjectsSection.tsx
@@ -3,7 +3,7 @@ import { mergeProjectsTextsAndMetadata } from '@/utils/projects.utils'
 import { projectsMetadata } from '@/constants/projects'
 import { i18n } from '@/constants/texts'
 
-import { SectionLayout } from '@/components/layout'
+import { SectionLayout } from '@/components/layouts'
 
 import ProjectList from '../ProjectList/ProjectList'
